Try to debug embedded bundle
