+++ package.json
@@ -127,6 +127,7 @@
     "@sourcegraph/prettierrc": "^3.0.3",
     "@sourcegraph/stylelint-config": "^1.4.0",
     "@sourcegraph/tsconfig": "^4.0.1",
+    "@statoscope/webpack-plugin": "^5.20.1",
     "@storybook/addon-a11y": "^6.3.12",
     "@storybook/addon-actions": "^6.3.12",
     "@storybook/addon-console": "^1.2.3",
